Project Manager/Designer
In this process, Bailee and I worked through Google docs, initially, to develop the narrative outline of the introduction. I uploaded the final version which included my edits, particularly in making the story active/present tense. We found this story to have a playful nature with flow, but did discover that the production of the elements we created in our minds is more time consuming and difficult. We spent a great deal of time with narrative, coding, and Github and did not have extra time to work with AI or explore ways to bring these ideas to life. 

In addition, the story map was uploaded and an asset list. I utilized this is the same manner I would a spotting list for sound design. 

Thank you and please let me know if you have any questions.